+++ core/ui/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/ui/ui_module.kt
@@ -1,7 +1,7 @@
 package de.tum.informatics.www1.artemis.native_app.core.ui
 
 import de.tum.informatics.www1.artemis.native_app.core.ui.remote_images.ArtemisImageProvider
-import de.tum.informatics.www1.artemis.native_app.core.ui.remote_images.ArtemisImageProviderImpl
+import de.tum.informatics.www1.artemis.native_app.core.ui.remote_images.impl.ArtemisImageProviderImpl
 import org.koin.dsl.module
 
 val uiModule = module {
